+++ Project1/Box.cpp
@@ -1,0 +1,47 @@
+#include "Box.h"
+#include <glut.h>
+
+void drawWireBoxWithoutTop() {
+  glBegin(GL_LINES);
+
+  // Bottom
+  glVertex3f(-1.0, -1.0, -1.0);
+  glVertex3f(1.0, -1.0, -1.0);
+
+  glVertex3f(1.0, -1.0, -1.0);
+  glVertex3f(1.0, 1.0, -1.0);
+
+  // glVertex3f(1.0, 1.0, -1.0);
+  // glVertex3f(-1.0, 1.0, -1.0);
+
+  glVertex3f(-1.0, 1.0, -1.0);
+  glVertex3f(-1.0, -1.0, -1.0);
+
+  // Top
+  glVertex3f(-1.0, -1.0, 1.0);
+  glVertex3f(1.0, -1.0, 1.0);
+
+  glVertex3f(1.0, -1.0, 1.0);
+  glVertex3f(1.0, 1.0, 1.0);
+
+  // glVertex3f(1.0, 1.0, 1.0);
+  // glVertex3f(-1.0, 1.0, 1.0);
+
+  glVertex3f(-1.0, 1.0, 1.0);
+  glVertex3f(-1.0, -1.0, 1.0);
+
+  // Sides
+  glVertex3f(-1.0, -1.0, -1.0);
+  glVertex3f(-1.0, -1.0, 1.0);
+
+  glVertex3f(1.0, -1.0, -1.0);
+  glVertex3f(1.0, -1.0, 1.0);
+
+  // glVertex3f(1.0, 1.0, -1.0);
+  // glVertex3f(1.0, 1.0, 1.0);
+
+  // glVertex3f(-1.0, 1.0, -1.0);
+  // glVertex3f(-1.0, 1.0, 1.0);
+
+  glEnd();
+}
