+++ Project1/Box.h
@@ -1,0 +1,7 @@
+#ifndef BOX_H
+#define BOX_H
+
+
+void drawWireBoxWithoutTop();
+
+#endif
