+++ Project1/Project1.vcxproj
@@ -131,10 +131,20 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Ball.cpp" />
+    <ClCompile Include="Box.cpp" />
+    <ClCompile Include="Collision.cpp" />
+    <ClCompile Include="Commands.cpp" />
     <ClCompile Include="Main.cpp" />
+    <ClCompile Include="Scenes.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Ball.h" />
+    <ClInclude Include="Box.h" />
+    <ClInclude Include="Collision.h" />
+    <ClInclude Include="Commands.h" />
     <ClInclude Include="include\glut.h" />
+    <ClInclude Include="Scenes.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
