+++ Project1/Scenes.h
@@ -1,0 +1,13 @@
+#ifndef SCENES_H
+#define SCENES_H
+
+#include <glut.h>
+
+#include "Ball.h"
+#include "Box.h"
+#include "Commands.h"
+#include "Collision.h"
+
+void drawMainScene();
+
+#endif
